Update the documented Ubuntu library dependencies
The docs currently list `libpcre3-dev` as a dependency, but this isn't actually used by ClamAV and will cause it to fallback to using the original PCRE library (and issue a warning saying that using libpcre2 is recommended). If the `libpcre2-dev` package is installed, ./configure picks that up and will use it.